Replace the build system with create-react-app
The main motivation for this is that while I like parcel a lot, v2 is
still in beta and has enough bugs that it gets in the way, and v1
doesn't have the features I'm looking for (mostly, it's the fast and
accurate HMR).

From Git's point of view, the biggest change here is that I'm switching
from "API and web client sharing the top level" to "web client at the
top level, and API in a subdirectory". This reflects the fact that the
user's experience is mostly about the front-end.