Better way to serve static files from the back-end
Rather than naming the files explicitly, we just add a handler after the
API routes that checks for the existence of the named file in the assets
directory. So the front-end can put whatever it likes in the build, and
as long as the paths don't conflict with the API routes (unlikely!),
we'll serve them up.